feat(aws): remove legacy unused s3 client
