+++ examples/CMakeLists.txt
@@ -11,10 +11,7 @@ cxx_link(echo_server base fibers2 http_server_lib TRDP::gperf)
 # cxx_link(proactor_stress base uring_fiber_lib http_server_lib)
 
 add_executable(s3_demo s3_demo.cc)
-cxx_link(s3_demo aws_lib)
-
-add_executable(s3v2_demo s3v2_demo.cc)
-cxx_link(s3v2_demo base aws_lib awsv2_lib)
+cxx_link(s3_demo base awsv2_lib)
 
 add_executable(https_client_cli https_client_cli.cc)
 cxx_link(https_client_cli base fibers2 http_client_lib tls_lib)
