+++ util/CMakeLists.txt
@@ -8,5 +8,4 @@ add_subdirectory(html)
 add_subdirectory(metrics)
 add_subdirectory(tls)
 add_subdirectory(http)
-add_subdirectory(cloud)
 add_subdirectory(aws)
